+++ discord_clicker/Models/Items/IItem.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using discord_clicker.Models.Items.AchievementClasses;
 using discord_clicker.Models.Person;
 using Microsoft.EntityFrameworkCore;
@@ -9,5 +10,5 @@ public interface IItem<TItem, TItemCreateModel> where TItem : class, IItem<TItem
     public int Id { get; set; }
     public string Name { get; set; }
     public string Description { get; set; }
-    public (bool, string, User) Get(User user, decimal money, DbSet<Achievement> achievements);
+    public (bool, string, User) Get(User user, long money, DbSet<Achievement> achievements);
 }
